requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
xgboost==2.1.4
shap==0.47.1
osmnx==2.0.2
networkx==3.4.2
httpx==0.28.1
python-dotenv==1.1.0
optuna==4.3.0
pandas==2.2.3
numpy==2.1.3
pydantic==2.11.2
scikit-learn==1.6.1
pyarrow==19.0.1
